refactor: move bench ensure funcs to compile packages
