+++ compile/ts/tools.go
@@ -1,0 +1,45 @@
+package tscode
+
+import (
+	"fmt"
+	"os"
+	"os/exec"
+	"path/filepath"
+)
+
+// EnsureDeno verifies that the Deno binary is installed and attempts to
+// install it if missing. It is safe to call from tests.
+func EnsureDeno() error {
+	return ensureDeno()
+}
+
+func ensureDeno() error {
+	if _, err := exec.LookPath("deno"); err == nil {
+		return nil
+	}
+	fmt.Println("\U0001F985 Installing Deno...")
+	home := os.Getenv("HOME")
+	if home == "" {
+		home = "/tmp"
+	}
+	installDir := filepath.Join(home, ".deno")
+	cmd := exec.Command("sh", "-c", fmt.Sprintf("curl -fsSL https://deno.land/install.sh | DENO_INSTALL=%s sh", installDir))
+	cmd.Stdout = os.Stdout
+	cmd.Stderr = os.Stderr
+	if err := cmd.Run(); err != nil {
+		return err
+	}
+	denoSrc := filepath.Join(installDir, "bin", "deno")
+	if _, err := os.Stat(denoSrc); err == nil {
+		if err := exec.Command("install", "-m", "755", denoSrc, "/usr/local/bin/deno").Run(); err == nil {
+			return nil
+		}
+		dest := filepath.Join(home, "bin", "deno")
+		if err := os.MkdirAll(filepath.Dir(dest), 0755); err == nil {
+			if err := os.Rename(denoSrc, dest); err == nil {
+				return nil
+			}
+		}
+	}
+	return fmt.Errorf("failed to install deno")
+}
